requirements.txt
RPi.GPIO
fake_rpi